Refactor files for DDD
