+++ src/Ad/model.js
@@ -1,0 +1,14 @@
+const Sequelize = require('sequelize');
+
+const Ad = {
+    id: {
+        type: Sequelize.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+    },
+    title: Sequelize.STRING,
+    description: Sequelize.TEXT,
+    availableSpots: Sequelize.INTEGER,
+};
+
+module.exports = Ad;
