--- src/controllers/utils.js
@@ -1,7 +1,0 @@
-exports.asyncHandler = handler => async (req, res, next) => {
-  try {
-    await handler(req, res, next);
-  } catch (err) {
-    next(err);
-  }
-};
