--- src/middleware/requireAuth.js
@@ -1,10 +1,0 @@
-const requireAuth = (req, res, next) => {
-  if (!req.user) {
-    const err = new Error('Forbidden')
-    err.status = 403
-    return next(err)
-  }
-  next()
-}
-
-module.exports = requireAuth
