+++ src/server.js
@@ -2,70 +2,20 @@ const express = require('express');
 const session = require('express-session');
 const bodyParser = require('body-parser');
 const morgan = require('morgan');
-const passport = require('passport');
 
+const { createFakeLoginSystem } = require('./User/authentication');
 const config = require('../config');
-const { db, Event, User, Ad } = require('./models');
+const connector = require('./connector');
+const { Event, Ad } = require('./connector');
 
-const events = require('./controllers/events');
+const isDev = config.get('env') === 'development';
 
 const app = express();
 
-const createFakeLoginSystem = () => {
-  const PassportStrategy = require('passport-strategy');
-  class LocalStrategy extends PassportStrategy {
-    constructor(verify) {
-      super();
-      this.name = 'local';
-      this._verify = verify;
-    }
-    authenticate(req) {
-      const email = req.body.email;
-      if (!email) {
-        return this.fail({ message: 'Missing Credentials' }, 400);
-      }
+app.set('env', config.get('env'));
+app.use(morgan(isDev ? 'dev' : 'common'));
+app.use(bodyParser.json());
 
-      this._verify(email)
-        .then(user => {
-          this.success(user);
-        })
-        .catch(err => {
-          this.error(err);
-        });
-    }
-  }
-
-  passport.use(
-    new LocalStrategy(email =>
-      User.findOrCreate({
-        where: { email },
-        defaults: { email },
-      }).spread(user => user),
-    ),
-  );
-
-  passport.serializeUser((user, done) => {
-    done(null, user.id);
-  });
-
-  passport.deserializeUser(async (id, done) => {
-    try {
-      done(null, await User.findById(id));
-    } catch (err) {
-      done(err);
-    }
-  });
-
-  app.use(passport.initialize());
-  app.use(passport.session());
-  app.post('/local-login', passport.authenticate('local'), (req, res) => {
-    res.send(req.user);
-  });
-};
-
-app.set('env', process.env.NODE_ENV || 'development');
-
-app.use(morgan(app.get('env') === 'development' ? 'dev' : 'common'));
 app.use(
   session({
     secret: config.get('session_secret'),
@@ -73,33 +23,29 @@ app.use(
     saveUninitialized: false,
   }),
 );
-app.use(bodyParser.json());
 
-if (app.get('env') === 'development') {
-  createFakeLoginSystem();
+if (isDev) {
+  createFakeLoginSystem(app);
 }
 
-app.use('/events', require('./controllers/events')(Event, Ad));
-app.use('/profile', require('./controllers/profile')());
+app.use('/events', require('./Event/router')(Event, Ad));
+app.use('/profile', require('./Profile/router')());
 
 app.use(function notFound(req, res, next) {
   const err = new Error('Not found');
   err.status = 404;
+
   next(err);
 });
+
 app.use(function errorHandler(err, req, res, next) {
-  const response =
-    app.get('env') === 'development'
-      ? {
-          message: err.message,
-          stack: err.stack,
-        }
-      : {};
+  const response = isDev ? err : {};
   res.status(err.status || 500).send(response);
 });
 
 const port = process.env.PORT || 3000;
-db
+
+connector
   .sync()
   .then(() => {
     app.listen(port, () => {
